Add Readme

+++ Day2/convert_km-to-miles.py
@@ -1,3 +1,3 @@
-miles = float(input("Enter miles: "))
+miles = int(input("Enter miles: "))
 kilometer = miles * 1.60934
 print("{} miles eaqauls {:.2f} kilometers".format(miles, kilometer))
